+++ client/components/index.js
@@ -3,16 +3,15 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './Navbar'
+export {default as Navbar} from './Nav-bar'
 export {default as UserHome} from './UserHome'
 export {default as HomePage} from './HomePage'
 export {default as SinglePageView} from './SinglePageView'
 export {default as About} from './About'
-export {default as Footer} from './Footer'
+export {default as Footer} from './Foot-er'
 export {default as Cart} from './Cart'
 export {default as AllProducts} from './AllProducts'
 export {default as SingleCategory} from './SingleCategory'
 export {default as CheckoutForm} from './CheckoutForm'
 export {default as OrderView} from './OrderView'
 export {Login, Signup} from './AuthForm'
-
